+++ includes/js/views/widgets/relaxer.js
@@ -1,4 +1,5 @@
-/*! Copyright 2017 - Ortus Solutions (Compiled: 17-04-2017) */
+/*! Copyright 2019 - Ortus Solutions (Compiled: 06-03-2019) */
+
 define([ "Backbone", "models/RelaxerHistory" ], function(Backbone, HistoryModel) {
     "use strict";
     var Relaxer = Backbone.View.extend({